Remove ineffective personalizations
In a few iterations i realized that the 500ms timer at the top of the
wpm animation serves as refresh frequency so therefore my other
modifications based on wpm did not work.

For earlier return to kyria logo, i left the display width change to
move back to logo when a fifth of the screen has zero bars.